refactor(ecmascript): reorganize large modules into focused components
Split monolithic files into smaller, focused modules for better maintainability:

**constant_evaluation/call_expr.rs** (892 lines → 4 modules):
- Split into array_functions, math_functions, string_methods, and global_functions
- Each module handles specific function type evaluation

**side_effects/expressions.rs** (781 lines → 8 modules):
- Split by expression type: literal, unary, binary, array, object, class, member, call
- Preserved all helper functions and moved to appropriate modules

**constant_evaluation/mod.rs** (541 lines → 7 modules):
- Extracted traits, binary, logical, unary, member, call, and identifier modules
- Maintained public API compatibility

Benefits:
- Single responsibility per module (30-150 lines each)
- Easier navigation and maintenance
- Better Claude Code compatibility
- No functional changes - all tests pass
