+++ crates/oxc_ecmascript/src/side_effects/expressions/unary_expressions.rs
@@ -1,0 +1,35 @@
+use oxc_ast::ast::*;
+
+use crate::to_primitive::ToPrimitive;
+
+use super::super::{MayHaveSideEffects, MayHaveSideEffectsContext};
+
+impl<'a> MayHaveSideEffects<'a> for UnaryExpression<'a> {
+    fn may_have_side_effects(&self, ctx: &impl MayHaveSideEffectsContext<'a>) -> bool {
+        match self.operator {
+            UnaryOperator::Delete => true,
+            UnaryOperator::Void | UnaryOperator::LogicalNot => {
+                self.argument.may_have_side_effects(ctx)
+            }
+            UnaryOperator::Typeof => {
+                if matches!(&self.argument, Expression::Identifier(_)) {
+                    false
+                } else {
+                    self.argument.may_have_side_effects(ctx)
+                }
+            }
+            UnaryOperator::UnaryPlus => {
+                // ToNumber throws an error when the argument is Symbol / BigInt / an object that
+                // returns Symbol or BigInt from ToPrimitive
+                self.argument.to_primitive(ctx).is_symbol_or_bigint() != Some(false)
+                    || self.argument.may_have_side_effects(ctx)
+            }
+            UnaryOperator::UnaryNegation | UnaryOperator::BitwiseNot => {
+                // ToNumeric throws an error when the argument is Symbol / an object that
+                // returns Symbol from ToPrimitive
+                self.argument.to_primitive(ctx).is_symbol() != Some(false)
+                    || self.argument.may_have_side_effects(ctx)
+            }
+        }
+    }
+}
